Refactor to wpx folder
