+++ packages/block-library/src/navigation/wpx/deep-signal.js
